Add support for custom types for ExpoPushMessage.data (Map<String, T>)
The default usage of PushClient and ExpoPush message remains.
The default type for the data member is now Map<String, Object>, which is closer to the
expo API documentation.
https://docs.expo.io/push-notifications/sending-notifications/#message-request-format
