+++ pom.xml
@@ -4,7 +4,7 @@
 
     <groupId>io.github.jav</groupId>
     <artifactId>expo-server-sdk</artifactId>
-    <version>0.7.1-SNAPSHOT</version>
+    <version>0.8.0-SNAPSHOT</version>
     <name>expo-server-sdk</name>
     <description>
         Java implementation of expo-server-sdk implementation.
